+++ StripeCore/StripeCore/Source/API Bindings/StripeAPI.swift
@@ -99,7 +99,7 @@ import PassKit
         return paymentRequest.paymentSummaryItems.last?.amount.floatValue ?? 0.0 >= 0
     }
 
-    class func supportedPKPaymentNetworks() -> [PKPaymentNetwork] {
+    @_spi(STP) public class func supportedPKPaymentNetworks() -> [PKPaymentNetwork] {
         return [
             .amex,
             .masterCard,