+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/Types/STPPaymentMethodCard.swift
@@ -34,6 +34,8 @@ public class STPPaymentMethodCard: NSObject, STPAPIResponseDecodable {
     @objc public private(set) var fingerprint: String?
     /// Contains information about card networks that can be used to process the payment.
     @objc public private(set) var networks: STPPaymentMethodCardNetworks?
+    /// Contains information about the card's display brand
+    @objc @_spi(STP) public private(set) var displayBrand: STPPaymentMethodDisplayBrand?
     /// Contains details on how this Card maybe be used for 3D Secure authentication.
     @objc public private(set) var threeDSecureUsage: STPPaymentMethodThreeDSecureUsage?
     /// If this Card is part of a Card Wallet, this contains the details of the Card Wallet.
@@ -62,6 +64,7 @@ public class STPPaymentMethodCard: NSObject, STPAPIResponseDecodable {
             "last4 = \(last4 ?? "")",
             "fingerprint = \(fingerprint ?? "")",
             "networks = \(networks?.description ?? "")",
+            "displayBrand = \(displayBrand?.description ?? "")",
             "threeDSecureUsage = \(threeDSecureUsage?.description ?? "")",
             "wallet = \(wallet?.description ?? "")",
         ]
@@ -89,6 +92,9 @@ public class STPPaymentMethodCard: NSObject, STPAPIResponseDecodable {
         card.fingerprint = dict.stp_string(forKey: "fingerprint")
         card.networks = STPPaymentMethodCardNetworks.decodedObject(
             fromAPIResponse: dict["networks"] as? [AnyHashable: Any]
+        )
+        card.displayBrand = STPPaymentMethodDisplayBrand.decodedObject(
+            fromAPIResponse: dict["display_brand"] as? [AnyHashable: Any]
         )
         card.threeDSecureUsage = STPPaymentMethodThreeDSecureUsage.decodedObject(
             fromAPIResponse: dict.stp_dictionary(forKey: "three_d_secure_usage"))